Changed: Apply Template pattern in `StringGenerator`
The classes extending `StringGenerator` has same form each other in `getLogString()`, so applied Template pattern for standardizing logging.
